Feature/social login 기능 일부 구현 (#9)
* feat: 네이버 소셜로그인 & 카카오 소셜로그인 기본 설정 (카카오톡 기본설정은 추후 예정)

* refactor: DB와 Column name 불일치 에러로 @Column 어노테이션으로 명시

* feat: 소셜로그인 및 JWT 토큰 관련 의존성 추가

* refactor: int -> long으로 타입 변경

* refactor: role을 문자열로 받기 위해 추가 설정

* refactor: id 칼럼 이름 명시, DTO로 값을 setter 이후 Entity에 수정하기 위한 메소드 추가

* feat: DB에서 회원 정보를 조회하는 메소드 추가

* feat: API 호출 시 데이터를 받아올 수 있는 DTO 생성

* feat: OAuth2 객체를 만들기 위한 DTO 생성

* feat: front-end와 연동을 위한 포트 설정

* feat: 인증을 위한 인증객체

* feat: 소셜로그인 후 정보를 받아 DB에 저장하고 Spring Security 필터체인에 쓰일 인증객체를 생성

* feat: 인증 성공 시 사용되는 Handler

* feat: Response 일관성을 위해 Handler 생성

* feat: JWT 필터를 통해서 토큰 검증

* feat: 토큰 관련 로직 클래스

* feat: Kakao와 Naver 소셜로그인이 쓰일 로직을 interface로 관리 (확장성)

* feat: Social 로그인 테스트용 Controller

* feat: Spring 인증 및 인가 관리 설정 클래스

* feat: Response 일관성을 위해 응답 객체 생성

+++ src/main/java/com/sparklenote/domain/entity/Paper.java
@@ -18,7 +18,7 @@ public class Paper extends BaseTimeEntity{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int paperId;
+    private Long paperId;
 
     private String content;
